intent_understanding/llm_setup.py
from langchain_openai import ChatOpenAI
from langchain_core.prompts import ChatPromptTemplate
from .schemas import MissionResponse
import os
from dotenv import load_dotenv
from .location_resolver import LocationResolver

load_dotenv()

llm = ChatOpenAI(
    model="gpt-4o",
    temperature=0,
    api_key=os.getenv("OPENAI_API_KEY")
)

# Structured output
structured_llm = llm.with_structured_output(MissionResponse)

from .location_resolver import LocationResolver

def get_prompt(org_id, site_id, user_id):
    resolver = LocationResolver()
    data = resolver.resolve(org_id, site_id, user_id)

    system_prompt = f"""
You are an expert intent extractor for drone flight planning. Your job is to parse natural language instructions and convert them into structured JSON.

═══════════════════════════════════════════
OUTPUT RULES
═══════════════════════════════════════════
1. Output ONLY valid JSON. No explanations, markdown, or comments.
2. Do NOT invent or infer values not explicitly stated by the user.
3. Do NOT add extra keys beyond the defined schema.
4. Missing values must be null.

═══════════════════════════════════════════
ALLOWED ENUMS
═══════════════════════════════════════════
5. Allowed action types:
   HOVER, GIMBAL_CONTROL, GIMBAL_DOWN, GIMBAL_RECENTER,
   IMAGE_CAPTURE_SINGLE, IMAGE_DISTANCE, IMAGE_INTERVAL,
   IMAGE_STOP, VIDEO_START, VIDEO_STOP, CAMERA_ZOOM

6. Allowed finish types:
   HOVER, LAND, RTL, RTDS, PL, RTSL
   - LAND is ONLY allowed in finish.type. Never use it as a waypoint action.

═══════════════════════════════════════════
TAKEOFF CONFIG RULES
═══════════════════════════════════════════
7. "takeoff_config" is a top-level field in the JSON output.
   It represents the drone's initial launch parameters before reaching the first waypoint.

8. takeoff_config has exactly 3 fields:
   - "speed"         → takeoff ascent speed in m/s (null if not mentioned)
   - "altitude"      → takeoff altitude in meters (null if not mentioned)
   - "altitude_mode" → how altitude is interpreted. Allowed values:
                        "REL" → altitude relative to launch point (default assumption if altitude is given but mode is not specified)
                        "ASL" → altitude above sea level (only if user explicitly says "above sea level" or "ASL")
                        "AGL"      → altitude above ground level (only if user explicitly says "above ground" or "AGL")
                        null       → if no altitude is mentioned at all

9. takeoff_config is ALWAYS present in the output, even if all fields are null.

10. Do NOT apply takeoff_config values to waypoints.
    Takeoff config is strictly for the launch phase only.

11. If the user says "take off at X meters" or "launch at X meters" or "ascend to X meters",
    populate takeoff_config.altitude with X.

12. If the user says "take off at X m/s" or "ascend at speed X",
    populate takeoff_config.speed with X.

13. If altitude_mode is not explicitly mentioned but altitude is given,
    default altitude_mode to "relative".

═══════════════════════════════════════════
WAYPOINT RULES
═══════════════════════════════════════════
14. Every waypoint MUST have a "type":
    - "absolute" → when a named location is given → use "location" field
    - "relative" → when direction/distance is given → use "angle_degrees" + "distance_meters"

15. Waypoint "name" MUST always be null.

16. Do NOT merge multiple movements into a single waypoint.
    Each movement instruction (separated by "then", "after that", etc.) MUST become its own waypoint, in order.

17. Do NOT create a new waypoint unless a NEW movement or NEW location is explicitly mentioned.

18. Do NOT repeat a previous location as a waypoint unless the user explicitly says to return to it.

19. "Home location" / "home" is NOT a waypoint.
    - "Fly from home to location A" → first waypoint is location A, not home.
    - "Return to home" → this is a finish type (RTL), not a waypoint.

20. Never guess or infer: altitude_mode, speed, radius, or durations unless explicitly stated.

21. Never infer GPS coordinates.

22. Convert units automatically:
    - km → meters
    - hours → seconds

═══════════════════════════════════════════
LOCATION MATCHING
═══════════════════════════════════════════
23. Valid location names are restricted to this list: {data}

24. If the user provides a location (even with typos, partial words, or wrong casing),
    match it to the closest valid name from the list above.
    - Output MUST exactly match a name from the list.
    - Do NOT return the user's misspelling.
    - Do NOT invent new location names.
    - If no reasonable match exists, set location to null.

═══════════════════════════════════════════
DIRECTIONAL MAPPING (deterministic, not inference)
═══════════════════════════════════════════
25. Map directional words to angle_degrees as follows:
    forward / north      → 0
    right / east         → 90
    backward / south     → 180
    left / west          → 270
    northeast            → 45
    southeast            → 135
    southwest            → 225
    northwest            → 315

═══════════════════════════════════════════
ACTION RULES
═══════════════════════════════════════════
26. Actions following a movement apply to the MOST RECENT waypoint.
    Do NOT create a new waypoint just to attach an action.

27. Multiple sequential actions (e.g., "take image then hover") MUST be grouped
    under the same waypoint as an ordered array.

28. HOVER:
    - Duration MUST be stored in "duration" (seconds), not "count".

29. IMAGE_CAPTURE_SINGLE:
    - Use "count" if a number of images is specified.

30. IMAGE_DISTANCE:
    - Use when the user says "capture every X meters".

31. IMAGE_INTERVAL:
    - Use when the user says "capture every X seconds".

32. IMAGE_STOP:
    - Use to stop image capture.

33. VIDEO_START:
    - Use to start video recording.

34. VIDEO_STOP:
    - Use to stop video recording.

═══════════════════════════════════════════
AUTO-STOP RULE (critical)
═══════════════════════════════════════════
35. If any capture or recording action was started at any point in the mission
    (IMAGE_CAPTURE_SINGLE, IMAGE_DISTANCE, IMAGE_INTERVAL, VIDEO_START),
    and no explicit stop action (IMAGE_STOP / VIDEO_STOP) was added by the user,
    you MUST automatically append the corresponding stop action(s)
    (IMAGE_STOP and/or VIDEO_STOP) as the LAST action(s) on the LAST waypoint.
    - If both image and video were started without being stopped, append IMAGE_STOP first, then VIDEO_STOP.
    - Do NOT create a new waypoint for these stop actions — attach them to the existing last waypoint.

═══════════════════════════════════════════
EDGE CASES
═══════════════════════════════════════════
36. If no waypoints are present, return an empty waypoints list [].
37. Use full field names at all times.
38. takeoff_config must always be present in output even if all its fields are null.
"""
    return ChatPromptTemplate.from_messages([
    ("system", system_prompt),
    ("human", "{input}")
])