requirements.txt
wordfreq
nltk
logger
logging
emoji
tiktoken

langchain
langchain_openai
pydantic
pydantic-settings
langchain_community
faiss-cpu
neo4j
python-socketio[asgi]
uvicorn
fastapi
mysql-connector-python



wordfreq==3.1.1
nltk==3.9.2
logger==1.4
logging==0.4.9.6
emoji==2.15.0
tiktoken==0.12.0
python-dotenv==1.2.1
langchain==1.2.6
langchain-openai==1.1.7
langchain-community==0.4.1

websocket-client==1.9.0
dotenv==0.9.9
faiss-cpu==1.13.2
fastapi==0.135.1
mysql-connector-python==9.6.0



marshmallow==3.26.2
msgpack==1.1.2
multidict==6.7.1
mypy_extensions==1.1.0
neo4j==6.1.0

numpy==2.4.1
openai==2.15.0
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
propcache==0.4.1
psycopg==3.3.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-engineio==4.13.1
python-socketio==5.16.1
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.52.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.41.0
wcwidth==0.2.14

wsproto==1.3.2
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0